Be more compatible with ansible_collections
