+++ playbooks/00-apiserver-proxy.yml
@@ -1,3 +1,1 @@
-- hosts: '{{ kube_cp_group|default("kube_control_plane") }}:{{ kube_worker_group|default("kube_workers") }}'
-  roles:
-    - role: apiserver_proxy
+00_apiserver_proxy.yml
