+++ playbooks/00_apiserver_proxy.yml
@@ -1,0 +1,3 @@
+- hosts: '{{ kube_cp_group|default("kube_control_plane") }}:{{ kube_worker_group|default("kube_workers") }}'
+  roles:
+    - role: apiserver_proxy
