+++ playbooks/00_inventory.yml
@@ -1,0 +1,5 @@
+---
+- hosts: all
+  gather_facts: false
+  roles:
+    - role: discover
